Create package structure for entities, enums, and utilities; move files accordingly.
