+++ src/main/java/org/example/enums/StadiumSectors.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.enums;
 
 public enum StadiumSectors {
     A,
